+++ src/views/PublicationCenter/PublicationCenter.ts
@@ -1,4 +1,11 @@
-import { type App, Modal, getIcon, TFile, Vault } from "obsidian";
+import {
+	type App,
+	Modal,
+	getIcon,
+	TFile,
+	ProgressBarComponent,
+	Vault,
+} from "obsidian";
 import QuartzSyncerSettings from "src/models/settings";
 import QuartzSyncerSiteManager from "src/repositoryConnection/QuartzSyncerSiteManager";
 import PublishStatusManager from "src/publisher/PublishStatusManager";
@@ -61,6 +68,18 @@ export class PublicationCenter {
 
 		return icon;
 	}
+
+	/**
+	 * Creates a progress bar component for displaying the publishing progress.
+	 * It returns a new instance of ProgressBarComponent.
+	 *
+	 * @returns A ProgressBarComponent instance.
+	 */
+	private progressBar = (element?: HTMLElement) => {
+		const containerEl = element ?? document.createElement("div");
+
+		return new ProgressBarComponent(containerEl);
+	};
 
 	/**
 	 * Shows the diff between the remote and local versions of a note.
@@ -132,9 +151,12 @@ export class PublicationCenter {
 				const diff = Diff.diffLines(remoteFile, localFile);
 				let diffView: DiffView | undefined;
 				const diffModal = new Modal(this.modal.app);
+				const title = notePath.split("/").pop() || "Diff";
 
 				diffModal.titleEl
-					.createEl("span", { text: `${notePath.split("/")[-1]}` })
+					.createEl("span", {
+						text: `${title}`,
+					})
 					.prepend(this.getIcon("file-diff"));
 
 				diffModal.onOpen = () => {
@@ -174,6 +196,7 @@ export class PublicationCenter {
 				props: {
 					publishStatusManager: this.publishStatusManager,
 					publisher: this.publisher,
+					progressBar: this.progressBar,
 					showDiff: this.showDiff,
 					close: () => {
 						this.modal.close();